remove warning

+++ src/linklayer/contract/MACAddress.h
@@ -136,12 +136,12 @@ class INET_API MACAddress
     /**
      * Sets the address to the broadcast address (hex ff:ff:ff:ff:ff:ff).
      */
-    void setBroadcast() {address = (macAddress64 ? MAC_ADDRESS_MASK64 : MAC_ADDRESS_MASK); }
+    void setBroadcast() {address = (macAddress64 ? (uint64_t)MAC_ADDRESS_MASK64 :(uint64_t)MAC_ADDRESS_MASK); }
 
     /**
      * Returns true if this is the broadcast address (hex ff:ff:ff:ff:ff:ff).
      */
-    bool isBroadcast() const {return (macAddress64 ? address == MAC_ADDRESS_MASK64 : address == MAC_ADDRESS_MASK); }
+    bool isBroadcast() const {return (macAddress64 ? address == (uint64_t)MAC_ADDRESS_MASK64 : address == MAC_ADDRESS_MASK); }
 
     /**
      * Returns true if this is a multicast logical address (first byte's lsb is 1).
@@ -214,7 +214,7 @@ class INET_API MACAddress
 
     MACAddress getMaskEUI48() { return MACAddress(MAC_ADDRESS_MASK); }
 
-    MACAddress getMaskEUI64() { return MACAddress(MAC_ADDRESS_MASK64); }
+    MACAddress getMaskEUI64() { return MACAddress((uint64_t)MAC_ADDRESS_MASK64); }
 
     bool getFlagEui64() const {return macAddress64; }
     void setFlagEui64(bool v)  {macAddress64 = v; }
